+++ openeo/rest/imagecollectionclient.py
@@ -27,13 +27,12 @@ class ImageCollectionClient(ImageCollection):
         Supports 0.4.
     """
 
-    def __init__(self, node_id: str, builder: GraphBuilder, session: 'Connection', metadata: CollectionMetadata=None):
-        super().__init__(metadata=metadata)
+    def __init__(self, node_id: str, builder: GraphBuilder, session: 'Connection', metadata: CollectionMetadata = None):
         self.node_id = node_id
         self.builder= builder
         self.session = session
         self.graph = builder.processes
-        self.metadata = metadata
+        self.metadata = CollectionMetadata.get_or_create(metadata)
 
     def __str__(self):
         return "ImageCollection: %s" % self.node_id
@@ -966,12 +965,6 @@ class ImageCollectionClient(ImageCollection):
             return graph_add_aggregate_process(with_read_vector)
         else:
             polygons = mapping(polygon)
-            polygons['crs'] = {
-                'type': 'name',
-                'properties': {
-                    'name': 'EPSG:4326'
-                }
-            }
             return graph_add_aggregate_process(self)
 
     def save_result(self, format: str = "GTIFF", options: dict = None):