+++ CHANGELOG.md
@@ -8,12 +8,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Make `DataCube.filter_bbox()` easier to use: allow passing a bbox tuple, list, dict or even shapely geometry directly as first positional argument or as `bbox` keyword argument. 
+  Handling of the legacy non-standard west-east-north-south positional argument order is preserved for now ([#136](https://github.com/Open-EO/openeo-python-client/issues/136))
+- Add "band math" methods `DataCube.ln()`, `DataCube.logarithm(base)`, `DataCube.log10()` and `DataCube.log2()`
+
 ### Changed
 
 ### Removed
 
 ### Fixed
 
+
+## [0.4.8] - 2020-11-17
+
+### Added
+- Add `DataCube.aggregate_spatial()`
+
+### Changed
+- Get/create default `RefreshTokenStore` lazily in `Connection`
+- Various documentation tweaks
 
 ## [0.4.7] - 2020-10-22